Add auto detect of product type
